fix: revert adding the Hook into its own set and move class
